Fix session user_id to set g.user
Change session['id'], when user logs in, the view sets
'user_id' instead of 'id'. Also, when there's no user logged
in trying to access session['id'] throws a key error
exception since the identifier is not present. Using
the .get method is the solution to prevent this exception
to happen.